+++ src/FormControls/index.ts
@@ -1,3 +1,3 @@
-export * from './GroupControl'
+export * from '../cdk/Controls/GroupControl'
 export * from './MapSelect'
 export * from './MapFilteredSelect'
